[chore] 기능 별 Model 및 Service 분류
