+++ app/src/main/java/com/example/sync_front/api_server/RetrofitClient.kt
@@ -1,18 +1,13 @@
 package com.example.sync_front.api_server
 
 import com.example.sync_front.BuildConfig
-import com.example.sync_front.data.service.HomeService
-import com.example.sync_front.data.service.SyncDetailService
+import com.example.sync_front.data.service.*
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
-import okhttp3.MediaType
 import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
 import java.text.SimpleDateFormat
 import java.util.*
