+++ app/src/main/java/com/example/sync_front/data/model/CategoryTypeModel.kt
@@ -1,15 +1,5 @@
-package com.example.sync_front.api_server
+package com.example.sync_front.data.model
 
-data class OnboardingModel(
-    val language: String,
-    val userName: String,
-    val countryName: String,
-    val gender: String,
-    val university: String,
-    val email: String,
-    val syncType: String,
-    val categoryTypes: CategoryTypes
-)
 
 data class CategoryTypes(
     val foreignLanguage: ForeignLanguage,
