+++ app/src/main/java/com/example/sync_front/data/model/CountriesModel.kt
@@ -1,0 +1,12 @@
+package com.example.sync_front.data.model
+
+data class CountriesRequestModel(
+    val page: String,
+    val perPage: String,
+    val language: String
+)
+data class CountriesResponse(
+    val status: Int,
+    val message: String,
+    val data: List<String>
+)
