+++ app/src/main/java/com/example/sync_front/data/service/CommunityService.kt
@@ -1,25 +1,10 @@
-package com.example.sync_front.api_server
+package com.example.sync_front.data.service
 
-import com.google.gson.annotations.SerializedName
+import com.example.sync_front.data.model.*
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Call
 import retrofit2.http.*
-
-interface LoginService {
-
-    // 소셜 로그인
-    @POST("auth/signin")
-    fun signIn(
-        @Header("Content-Type") application: String,
-        @Header("Authorization") accessToken: String,
-        @Header("fcmToken") fcmToken: String,
-        @Body platform: Platform
-    ): Call<LogInResponse>
-
-    // 회원 가입
-
-}
 
 interface CommunityService {
     //게시글 조회
@@ -95,39 +80,4 @@ interface CommunityService {
         @Header("Authorization") accessToken: String,
         @Path("commentId") commentId: Int
     ): Call<Void>
-}
-
-interface GoogleService {
-    // 구글 액세스 토큰 발급
-    @POST("oauth2/v4/token")
-    fun getAccessToken(
-        @Body request: LoginGoogleRequestModel
-    ): Call<LoginGoogleResponseModel>
-}
-
-interface CountriesService { //국가 조회
-    @POST("auth/countries")
-    fun getCountries(
-        @Header("Content-Type") application: String,
-        @Body request: CountriesRequestModel
-    ): Call<CountriesResponse>
-}
-
-interface EmailService { // 학교 이메일 인증
-    @POST("auth/school-emails/verification-requests")
-    fun sendEmail(
-        @Header("Content-Type") application: String,
-        @Body request: EmailRequest
-    ): Call<EmailResponse>
-
-    @POST("auth/school-emails/verifications")
-    fun sendCode(
-        @Header("Content-Type") application: String,
-        @Body request: CodeRequest
-    ): Call<CodeResponse>
-
-    @POST("auth/school-emails/reset")
-    fun sendReset(
-        @Header("Content-Type") application: String
-    ): Call<CodeResetResponse>
 }
