+++ app/src/main/java/com/example/sync_front/data/service/CountriesService.kt
@@ -1,0 +1,16 @@
+package com.example.sync_front.data.service
+
+import com.example.sync_front.data.model.CountriesRequestModel
+import com.example.sync_front.data.model.CountriesResponse
+import retrofit2.Call
+import retrofit2.http.Body
+import retrofit2.http.Header
+import retrofit2.http.POST
+
+interface CountriesService { //국가 조회
+    @POST("auth/countries")
+    fun getCountries(
+        @Header("Content-Type") application: String,
+        @Body request: CountriesRequestModel
+    ): Call<CountriesResponse>
+}
